Fixed 101 103

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -8,7 +8,7 @@
 void print_number(int n)
 {
 	unsigned int tens, digit, intPos = n;
-	unsigned int digitNum = 1;
+	double digitNum = 1;
 
 	if (n == 0)
 		_putchar('0');
